fix a four-byte write into a one-byte coercion flag
qio_scatter_int64() still declared `int *coerced` after the per-column
warning change converted the flags to uint8_t. Callers pass a one-byte slot
in the per-column array, so `*coerced = 1` wrote four bytes: three of them
over the following columns' flags, or past the end of the allocation for the
last selected column.

Introduced today by the per-column warning commit, whose rename matched
`int *int64_coerced` and `int *sentinel` but not `int *coerced`.

The tests could not catch it -- the surviving three bytes landed in padding
or in flags that had already been read -- but the compiler warns about it,
and R CMD check reports the warning.

+++ src/qio_file.c
@@ -642,7 +642,7 @@ static void qio_place_int64_na(double *dst, int mode) {
 static void qio_scatter_int64(double *dst, const void *values,
                               const int16_t *def_levels, int16_t max_def,
                               int64_t length, int mode, int is_unsigned,
-                              int *coerced) {
+                              uint8_t *coerced) {
     const int64_t *src = (const int64_t *)values;
     int any = 0;
     if (def_levels == NULL) {
